requirements.txt
Django==4.2.7
web3==6.11.3
py-solc-x==2.0.2
python-dotenv==1.0.0
Pillow>=10.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
pyotp==2.9.0
qrcode==7.4.2
